requirements.txt
qrcode==7.4.2
pillow==10.2.0
pygame==2.6.1